Made name optional from Google people API
The Google people API might not return any names (usually due to privacy
configuration), but may still return the current avatar for the users
profile. This ensures that we don't fail to parse the returned JSON in
such cases, and use our existing record of their name if no name is
sent.